requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
idna==2.8
pycparser==2.19
PyJWT==1.7.1
requests==2.22.0
six==1.13.0
urllib3==1.25.7
